+++ lib/traktor_d2.ts
@@ -14,9 +14,10 @@
 
 import { BaseController } from "./base_controller";
 import d2Config from "./traktor_d2_config.json";
+import type { HidAdapterFactory, UsbAdapterFactory } from "./usb/adapter";
 
 export class TraktorD2 extends BaseController {
-  constructor() {
-    super(d2Config);
+  constructor(createHid: HidAdapterFactory, createUsb: UsbAdapterFactory) {
+    super(d2Config, createHid, createUsb);
   }
 }
