+++ lib/traktor_f1.ts
@@ -1,8 +1,9 @@
 import { BaseController } from "./base_controller";
 import f1Config from "./traktor_f1_config.json";
+import type { HidAdapterFactory, UsbAdapterFactory } from "./usb/adapter";
 
 export class TraktorF1 extends BaseController {
-  constructor() {
-    super(f1Config);
+  constructor(createHid: HidAdapterFactory, createUsb: UsbAdapterFactory) {
+    super(f1Config, createHid, createUsb);
   }
 }
