+++ tests/playground.ts
@@ -1,6 +1,7 @@
 import { MaschineMk3 } from "../lib/maschine_mk3";
+import { createNodeHidAdapter, createNodeUsbAdapter } from "../lib/usb/node";
 
-const maschineMk3 = new MaschineMk3();
+const maschineMk3 = new MaschineMk3(createNodeHidAdapter, createNodeUsbAdapter);
 
 maschineMk3
   .init()
